change js to jsx extensions and rename indexes
